Extract headers from go sources to h files
