+++ libs/glob.go
@@ -1,6 +1,8 @@
 package libs
 
 import (
+	_ "embed"
+
 	"github.com/gotranspile/cxgo/runtime/stdio"
 	"github.com/gotranspile/cxgo/types"
 )
@@ -8,6 +10,9 @@ import (
 const (
 	globH = "glob.h"
 )
+
+//go:embed glob.h
+var hglob string
 
 func init() {
 	RegisterLibrary(globH, func(c *Env) *Library {
@@ -34,22 +39,7 @@ func init() {
 				"GLOB_NOESCAPE": c.NewIdent("GLOB_NOESCAPE", "stdio.GlobNoEscape", stdio.GlobNoEscape, gint),
 			},
 			// TODO
-			Header: `
-#include <` + stddefH + `>
-
-const _cxgo_go_int GLOB_NOESCAPE = 1;
-
-typedef struct {
-    size_t   gl_pathc;    /* Count of paths matched so far  */
-    char   **gl_pathv;    /* List of matched pathnames.  */
-    size_t   gl_offs;     /* Slots to reserve in gl_pathv.  */
-	_cxgo_sint32 (*Glob)(const char *pattern, _cxgo_sint32 flags,
-                _cxgo_sint32 (*errfunc) (const char *epath, _cxgo_sint32 eerrno));
-	void (*Free)(void);
-} glob_t;
-#define glob(pattern, flags, errfunc, g) ((glob_t*)g)->Glob(pattern, flags, errfunc)
-#define globfree(g) ((glob_t*)g)->Free()
-`,
+			Header: hglob,
 		}
 	})
 }
